[MM-65728] Block notifications from immediate bot replies in threads (#390)
Currently when you message an agent, the agent replies to your post in a thread and always triggers an immediate notification. With this commit, I've introduced a mechanism that blocks notifications related to threaded bot replies. This will silence the immediate notifications you get when interacting with agents.

For web and desktop notifications, we block all threaded bot replies that are created within 1 second of the parent post. Note that this mechanism will block notifications for all bot posts with this behavior. **So if a bot replies immediately to a top-level user post and creates a thread, you will never get a web/desktop notification for that post.** We could introduce an API to provide all known agent bot names to the webapp and filter on that list, but I'm not sure it's worth the extra complexity.

For push and email notifications, we block all threaded bot replies that are created by known AI Agent bots (with no time restriction). **This means you will never receive a push or email notification from an AI Agent bot if that post is created within a thread.**